Move parquet ddl tests to own file
